+++ integrations/whatsapp/src/channels/channel.ts
@@ -14,6 +14,7 @@ import {
 import { getAuthenticatedWhatsappClient } from '../auth'
 import { WHATSAPP } from '../misc/constants'
 import { convertMarkdownToWhatsApp } from '../misc/markdown-to-whatsapp-rtf'
+import { splitTextMessageIfNeeded } from '../misc/split-text-message'
 import { sleep } from '../misc/util'
 import { repeat } from '../repeat'
 import * as card from './message-types/card'
@@ -255,24 +256,6 @@ function backoffDelayMs(attempt: number) {
 }
 
 const MAX_ATTEMPT = 3
-const WHATSAPP_MAX_TEXT_LENGTH = 4096
-
-export function splitTextMessageIfNeeded(message: string): OutgoingMessage[] {
-  const textLength = message.length
-  if (textLength <= WHATSAPP_MAX_TEXT_LENGTH) {
-    return [new Text(message)]
-  }
-
-  const numChunks = Math.ceil(textLength / WHATSAPP_MAX_TEXT_LENGTH)
-  const chunks = Array.from({ length: numChunks }, (_, i) => {
-    const start = i * WHATSAPP_MAX_TEXT_LENGTH
-    const end = (i + 1) * WHATSAPP_MAX_TEXT_LENGTH
-    const chunk = new Text(message.slice(start, end))
-    return chunk
-  })
-
-  return chunks
-}
 
 async function _send({ client, ctx, conversation, logger, message, ack }: SendMessageProps) {
   if (!message) {
